+++ rollup.config.js
@@ -1,7 +1,11 @@
 import typescript from "rollup-plugin-typescript2";
 import copy from "rollup-plugin-copy";
-import getTerser from "./rollup-plugins/terser.js";
+import getTerser from "./rollup-plugins/getTerser.js";
 import replace from "@rollup/plugin-replace";
+import nodeResolve from "@rollup/plugin-node-resolve";
+import commonjs from "@rollup/plugin-commonjs";
+import { getSass } from "./rollup-plugins/getSass.js";
+import path from "path";
 
 const isProduction = process.env.NODE_ENV === "production";
 const outputDir = isProduction ? "build-output" : "watch-output";
@@ -14,6 +18,7 @@ export default [
       dir: outputDir,
       format: "cjs",
       entryFileNames: "code/content/[name].js",
+      sourcemap: !isProduction,
     },
     watch: {
       clearScreen: false,
@@ -24,22 +29,49 @@ export default [
         targets: [{ src: ["src/**/*", "!**/*/code"], dest: outputDir }],
         flatten: false,
       }),
-      typescript(),
+      typescript({ abortOnError: isProduction }),
       replace({
         "process.env.NODE_ENV": JSON.stringify(process.env.NODE_ENV),
         preventAssignment: true,
       }),
       isProduction && terser,
+      getSass({
+        output: path.join(outputDir, "code", "content", "styles.css"),
+        isProduction,
+      }),
     ],
   },
   {
     input: "src/code/popup/index.ts",
     output: {
       dir: outputDir,
-      format: "cjs",
+      format: "iife",
       entryFileNames: "code/popup/[name].js",
+      sourcemap: !isProduction,
     },
-    plugins: [typescript(), isProduction && terser],
+    plugins: [
+      /* replace `process` in raw React bundle */
+      replace({
+        preventAssignment: true,
+        values: {
+          "process.env.NODE_ENV": JSON.stringify(
+            isProduction ? "production" : "development",
+          ),
+          "typeof process": '"undefined"',
+          process: "undefined",
+        },
+      }),
+      nodeResolve({
+        browser: true /* Resolves import paths such as that of `react` or `lodash` to node_modules/react/index.js so rollup can locate imported package */,
+      }),
+      commonjs() /* Converts CommonJS modules in React to ES6 modules */,
+      typescript({ abortOnError: isProduction }),
+      isProduction && terser,
+      getSass({
+        output: path.join(outputDir, "code", "popup", "styles.css"),
+        isProduction,
+      }),
+    ],
   },
   {
     input: "src/code/background/index.ts",
@@ -47,7 +79,11 @@ export default [
       dir: outputDir,
       format: "cjs",
       entryFileNames: "code/background/[name].js",
+      sourcemap: !isProduction,
     },
-    plugins: [typescript(), isProduction && terser],
+    plugins: [
+      typescript({ abortOnError: isProduction }),
+      isProduction && terser,
+    ],
   },
 ];