+++ src/code/background/index.ts
@@ -1,15 +1,26 @@
-import { addBrowserStorageListener, update } from "../browser-api";
+import {
+  addBrowserStorageListener,
+  queryActiveTab,
+  updateExtensionIcon,
+  updateTabContent,
+} from "../browser-api";
 import { debounceUpdate } from "./utils";
 import { BrowserEvents } from "../content/events";
 import { StateChanges } from "../browser-api/types";
 
-console.info("background script running");
+addBrowserStorageListener("onChanged", async (changes: StateChanges) => {
+  if (changes.extensionIsEnabled) {
+    void updateExtensionIcon({
+      extensionIsEnabled: changes.extensionIsEnabled.newValue,
+    });
+  }
+});
 
-// Update when the active tab changes
+/* when user switches to new tab */
 browser.tabs.onActivated.addListener(async (info) => {
-  console.info("browser.tabs.onActivated, info:", info);
-  void update({
+  void updateTabContent({
     browserEvent: BrowserEvents.TabsOnActivated,
+    previousTabId: info.previousTabId,
   });
 });
 
@@ -18,20 +29,38 @@ const getDebouncedTab = debounceUpdate(100);
 browser.tabs.onUpdated.addListener(async (_tabId, changeInfo, tab) => {
   if (changeInfo.status === "complete" && tab.active) {
     const lastTab = await getDebouncedTab(tab);
-    console.info("browser.tabs.onUpdated, tab:", lastTab, changeInfo);
-    void update({
+    void updateTabContent({
       browserEvent: BrowserEvents.TabsOnUpdated,
       activeTab: lastTab,
     });
   }
 });
 
-// Update when the storage state changes
-addBrowserStorageListener("onChanged", async (changes: StateChanges) => {
-  if (changes.extensionIsEnabled) {
-    void update({
-      browserEvent: BrowserEvents.StorageOnChanged,
-      extensionIsEnabled: changes.extensionIsEnabled.newValue,
-    });
-  }
+/*
+ * Only one `onMessage` listener can be added, as non-relevant listeners will return `undefined` to `sendMessage` call if they fire first
+ * gets tab ids and sends them back to content script
+ * `sendResponse` is not used as it requires a callback argument in `browser.runtime.sendMessage`
+ */
+browser.runtime.onMessage.addListener(
+  async (message: { action: "getTabIds" }, sender, _sendResponse) => {
+    if (message.action !== "getTabIds") {
+      throw new Error("actions other than `getTabIds` are not allowed");
+    }
+
+    const activeTab = await queryActiveTab();
+    const activeTabId = activeTab.id;
+
+    if (!sender.tab) {
+      console.info("getTabIds: No tab object exists on sender");
+      return { activeTabId };
+    }
+
+    return { contentTabId: sender.tab.id, activeTabId };
+  },
+);
+
+browser.management.onInstalled.addListener((_info) => {
+  void updateTabContent({
+    browserEvent: BrowserEvents.ManagementOnInstalled,
+  });
 });