+++ src/code/content/events.ts
@@ -1,6 +1,7 @@
 export enum BrowserEvents {
-  TabsOnActivated = "browser.tabs.OnActivated",
-  TabsOnUpdated = "browser.tabs.OnUpdated",
+  ManagementOnInstalled = "browser.management.onInstalled" /* extension has been installed or reloaded */,
+  TabsOnActivated = "browser.tabs.OnActivated" /* user switched to new tab */,
+  TabsOnUpdated = "browser.tabs.OnUpdated" /* url changed in the same tab, with or without page reload */,
   StorageOnChanged = "browser.storage.onChanged",
 }