+++ src/code/content/filter.ts
@@ -3,10 +3,12 @@ import Observer, {
   EmittedNodeEventHandler,
   ObservedElementDetail,
 } from "./observer";
-import { YouTubePageTypes } from "./types";
 import Element from "./element";
-import { MessagePayload } from "../browser-api/types";
+import { MessageToContentPayload, StateChanges } from "../browser-api/types";
 import {
+  addBrowserStorageListener,
+  getTabIds,
+  isActiveTab,
   isExtensionEnabled,
   updateHiddenVideoCount,
   updateVideoCount,
@@ -15,17 +17,13 @@ import { debounce } from "../utils";
 import { selectors } from "./selectors";
 import YouTube from "../utils/youtube";
 import Dom from "./dom";
-
-export interface Options {
-  watchedFilterEnabled: boolean;
-  membersOnlyFilterEnabled: boolean;
-  showVideoNumbers: boolean;
-}
+import { Filters, Options } from "../types";
 
 export default class Filter {
-  watchedFilterEnabled: boolean = true;
-  membersOnlyFilterEnabled: boolean = true;
-  videoElementsHidden = new Set();
+  watchedFilterEnabled = true;
+  membersOnlyFilterEnabled = false;
+  videoNumbersAreShown = false;
+  hiddenVideos = new Set();
   youTube = new YouTube();
   element = new Element();
   updateVideoCount = debounce(() => {
@@ -33,23 +31,23 @@ export default class Filter {
     void updateHiddenVideoCount(this.element.hiddenVideoCount);
   }, 100);
   numberVideos = debounce(async () => {
-    if (
-      (await isExtensionEnabled()) &&
-      this.options.showVideoNumbers &&
-      process.env.NODE_ENV === "development"
-    ) {
-      this.options.showVideoNumbers && this.element.numberVideos();
+    if (this.videoNumbersAreShown && (await isExtensionEnabled())) {
+      this.element.numberVideos();
     }
   }, 100);
 
   cleanUpProcedures: (() => void)[] = [];
 
-  public constructor(public options: Options) {
-    this.watchedFilterEnabled = this.options.watchedFilterEnabled;
-    this.membersOnlyFilterEnabled = this.options.membersOnlyFilterEnabled;
+  public constructor(filters: Filters, options: Options) {
+    const { watched, membersOnly } = filters;
+    const { videoNumbersAreShown } = options;
+    this.watchedFilterEnabled = watched;
+    this.membersOnlyFilterEnabled = membersOnly;
+    this.videoNumbersAreShown = videoNumbersAreShown;
 
     this.addYtNavigateFinishEventListener();
-    this.establishCommunicationWithBackground();
+    this.listenForMessageFromBackground();
+    this.listenForBrowserStorageChanges();
   }
 
   get youTubePageType() {
@@ -75,10 +73,10 @@ export default class Filter {
     }
   }
 
-  private async run(pageType: YouTubePageTypes) {
-    console.info(`*** Running filter on ${pageType} ***`);
+  private async run() {
+    this.element.pageType = this.youTubePageType;
+    console.info(`*** Running filter on ${this.element.pageType} ***`);
 
-    this.element.pageType = pageType;
     const contents = await this.element.waitForAndGetAndSetContents();
     const { videoElementTagName } = this.element;
     this.addContentsObserver(contents, videoElementTagName);
@@ -86,17 +84,13 @@ export default class Filter {
 
     this.cleanUpProcedures.push(
       () => {
-        this.videoElementsHidden.clear();
         console.info(
-          `Cleared the Set of ${this.videoElementsHidden.size} hidden video elements`,
+          `Cleared the Set of ${this.hiddenVideos.size} hidden video elements`,
         );
+        this.hiddenVideos.clear();
       },
       () => {
-        if (
-          this.options.showVideoNumbers &&
-          process.env.NODE_ENV === "development" &&
-          contents
-        ) {
+        if (contents) {
           const unNumberedVideoCount = this.element.unNumberVideos();
           console.info(
             `Removed ${unNumberedVideoCount} ${selectors.videoNumberClass} elements from the DOM`,
@@ -160,7 +154,7 @@ export default class Filter {
 
         if (
           !videoElement ||
-          this.videoElementsHidden.has(videoElement) ||
+          this.hiddenVideos.has(videoElement) ||
           !this.shouldHideVideo(triggeringElement)
         ) {
           return;
@@ -174,7 +168,7 @@ export default class Filter {
         );
 
         if (triggeringElementIsVideo) {
-          this.videoElementsHidden.delete(triggeringElement);
+          this.hiddenVideos.delete(triggeringElement);
         }
         break;
       default:
@@ -196,7 +190,7 @@ export default class Filter {
 
   private hideVideo(video: HTMLElement) {
     Dom.setVisibility(video, true);
-    this.videoElementsHidden.add(video);
+    this.hiddenVideos.add(video);
   }
 
   private showVideo(video: HTMLElement) {
@@ -204,8 +198,15 @@ export default class Filter {
   }
 
   private async filterLoadedVideos() {
+    const extensionIsEnabled = await isExtensionEnabled();
+
     for (const video of this.element.videos) {
-      if ((await isExtensionEnabled()) && this.shouldHideVideo(video)) {
+      if (!extensionIsEnabled) {
+        this.showVideo(video);
+        continue;
+      }
+
+      if (this.shouldHideVideo(video)) {
         this.hideVideo(video);
       } else {
         this.showVideo(video);
@@ -213,7 +214,10 @@ export default class Filter {
     }
 
     this.updateVideoCount();
-    this.numberVideos();
+
+    if (extensionIsEnabled && this.videoNumbersAreShown) {
+      this.element.numberVideos();
+    }
   }
 
   private addChipsListener() {
@@ -224,7 +228,7 @@ export default class Filter {
         this.cleanUp();
 
         if (await isExtensionEnabled()) {
-          void this.run(this.youTubePageType);
+          void this.run();
         }
       }
     };
@@ -250,7 +254,7 @@ export default class Filter {
 
       try {
         if (await isExtensionEnabled()) {
-          void this.run(this.youTubePageType);
+          void this.run();
         }
       } catch (e) {
         console.info(
@@ -262,49 +266,87 @@ export default class Filter {
     window.addEventListener(youTubeEvents.ytNavigateFinish, handler);
   }
 
-  private establishCommunicationWithBackground() {
-    // Listen for toggle messages initiated by popup and by background script, which are sent while on YouTubePageTypes pages only
+  private listenForBrowserStorageChanges() {
+    addBrowserStorageListener("onChanged", async (changes: StateChanges) => {
+      const { extensionIsEnabled } = changes;
+
+      if (!(extensionIsEnabled && (await isActiveTab()))) return;
+
+      if (extensionIsEnabled.newValue) {
+        this.cleanUp();
+        void this.run();
+      } else {
+        await this.filterLoadedVideos();
+        this.cleanUp();
+      }
+    });
+
+    addBrowserStorageListener("onChanged", async (changes: StateChanges) => {
+      const { filters } = changes;
+
+      if (!(filters && (await isActiveTab()))) return;
+
+      this.watchedFilterEnabled = filters.newValue.watched;
+      this.membersOnlyFilterEnabled = filters.newValue.membersOnly;
+
+      await this.filterLoadedVideos();
+    });
+
+    addBrowserStorageListener("onChanged", async (changes: StateChanges) => {
+      const { options } = changes;
+
+      if (!(options && (await isActiveTab()))) return;
+
+      this.videoNumbersAreShown = options.newValue.videoNumbersAreShown;
+
+      if (this.videoNumbersAreShown) {
+        this.element.numberVideos();
+      } else {
+        this.element.unNumberVideos();
+      }
+    });
+  }
+
+  private listenForMessageFromBackground() {
     browser.runtime.onMessage.addListener(
-      async (message: MessagePayload, _sender, _sendResponse) => {
-        const {
-          tabId,
-          extensionIsEnabled,
-          browserEvent,
-          currentYouTubePageType,
-        } = message;
-
-        console.info("tabId:", tabId);
-        console.info("browser event:", browserEvent);
+      async (message: MessageToContentPayload, _sender, _sendResponse) => {
+        const { activeTabId, browserEvent } = message;
 
         if (
-          currentYouTubePageType &&
-          extensionIsEnabled &&
-          [
-            BrowserEvents.StorageOnChanged /* extension just toggled on */,
-            BrowserEvents.TabsOnActivated /* navigated to new tab */,
-          ].includes(browserEvent)
+          ![
+            BrowserEvents.TabsOnActivated,
+            BrowserEvents.ManagementOnInstalled,
+          ].includes(browserEvent) ||
+          activeTabId === undefined
         ) {
-          this.cleanUp();
-          void this.run(currentYouTubePageType);
-
           return;
         }
 
-        if (
-          extensionIsEnabled &&
-          BrowserEvents.TabsOnUpdated === browserEvent
-        ) {
-          /* BrowserEvents.TabsOnUpdated debounced can fire even after youTubeEvents.ytNavigateFinish, on full page reload, which makes it useless for handling any logic */
+        const { contentTabId } = await getTabIds();
+        const isActiveTab = activeTabId === contentTabId;
 
+        if (isActiveTab) {
+          void this.run();
+        }
+      },
+    );
+
+    browser.runtime.onMessage.addListener(
+      async (message: MessageToContentPayload, _sender, _sendResponse) => {
+        const { previousTabId, browserEvent } = message;
+
+        if (
+          BrowserEvents.TabsOnActivated !== browserEvent ||
+          previousTabId === undefined
+        ) {
           return;
         }
 
-        if (
-          /* extension just toggled off */
-          !extensionIsEnabled &&
-          BrowserEvents.StorageOnChanged === browserEvent
-        ) {
-          await this.filterLoadedVideos();
+        const { contentTabId } = await getTabIds();
+
+        const isPreviousTab = previousTabId === contentTabId;
+
+        if (isPreviousTab) {
           this.cleanUp();
         }
       },