+++ src/code/popup/selectors.ts
@@ -1,4 +1,0 @@
-export const selectors = {
-  toggleFilterButtonId: "toggle-filter",
-  extensionShouldNotRunElementId: "extension-should-not-run",
-};