+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2025-10-18
+
+### Added
+
+- Hide members-only videos
+- Show video numbers
+- Extension is auto enabled when installed
+
 ## [1.0.0] - 2025-10-01
 
 ### Added